added ConnectionManager
