+++ src/main/java/com/company/watches/dao/VendorDAO.java
@@ -7,7 +7,7 @@ import java.sql.*;
 import java.util.List;
 import java.util.Optional;
 
-import static com.company.watches.utils.Settings.getConnection;
+import static com.company.watches.utils.ConnectionManager.getConnection;
 
 class VendorDAO implements DAO<Vendor> {
